Created separate file to house HSGWO class
