+++ MultiAgentEnvironment.py
@@ -6,12 +6,14 @@ import math
 from matplotlib import pyplot as plt
 
 from maddpg_code import *
+from hsgwo_code import *
 import statistics
 import numpy
 import imageio
 import time
 import csv
 import pandas as pd
+
 
 
 # Constants
@@ -66,152 +68,6 @@ class Agent:
         self.x, self.y = self.initial_pos
 
 
-class Wolf(Agent):
-    def __init__(self, agent_id, x, y):
-        super().__init__(agent_id, x, y)
-        self.fitness = 0.0
-        self.search_radius = SEARCH_RADIUS
-        self.is_alpha = False
-        self.is_commensal = True
-        self.temppath = []
-
-    def heuristic(self, point):
-        dx = point[0] - self.x
-        dy = point[1] - self.y
-        return math.sqrt((dx ** 2 + dy ** 2))
-
-    def heuristic2(self, point, goal):
-        dx = point[0] - goal[0]
-        dy = point[1] - goal[1]
-        return math.sqrt((dx ** 2 + dy ** 2))
-
-    def make_alpha(self):
-        self.is_alpha = True
-
-    def make_omega(self):
-        self.is_alpha = False
-
-    def make_commensal(self):
-        self.is_commensal = True
-
-    def i_already_explored(self):
-        self.is_commensal = False
-
-    def is_valid(self, node, obstacles):
-        x, y = node
-        if x < 0 or x >= WIDTH or y < 0 or y >= HEIGHT:
-            return False
-        for obstacle in obstacles:
-            if ((x - obstacle.x) ** 2 + (y - obstacle.y) ** 2) ** 0.5 <= AGENT_RADIUS + obstacle.radius:
-                return False
-        return True
-
-    def is_visible(self, obstacle):
-        distance_from_obstacle = (((self.x - obstacle.x)**2 + (self.y - obstacle.y)**2)**0.5)-obstacle.radius
-        if distance_from_obstacle < (self.search_radius + obstacle.radius):
-            return True, distance_from_obstacle
-        else:
-            return False, distance_from_obstacle
-
-    def obstacles_in_range(self, obstacles):
-        # finds obstacles within search radius and returns visible obstacles and their distances
-        list_of_threats = []
-        for obstacle in obstacles:
-            threat, distance_from_obstacle = self.is_visible(obstacle)
-            if threat:
-                list_of_threats.append(1/(distance_from_obstacle**2))
-        return list_of_threats
-
-    def explore(self, goal, obstacles):
-        i = 0
-        j = 1
-        while self.is_commensal and i < 6:
-            # randomly generate an angle
-            new_angle = random.uniform(0, 2*math.pi)
-            # find a point on that angle
-            new_destination = (self.x + (np.cos(new_angle)*j*MOVEMENT_SPEED), self.y + (np.sin(new_angle)*j*MOVEMENT_SPEED))
-            value_new_destination = self.heuristic2(new_destination, goal)   # check the heuristic value of that point
-            i += 1
-            if value_new_destination < self.heuristic2((self.x, self.y), goal) and self.is_valid(new_destination, obstacles):
-                self.x = new_destination[0]
-                self.y = new_destination[1]
-                self.path.append((self.x, self.y))
-                self.temppath.append((self.x, self.y,))
-                self.i_already_explored()
-
-
-    def update_fitness(self, goal, obstacles):
-        # J_fuel = len(self.temppath)
-        J_threat = 0.0
-        mu = 0.2    # we liked 0.9
-        k = 500
-        list_of_threats = self.obstacles_in_range(obstacles)
-        for obstacle in list_of_threats:
-            J_threat = J_threat + obstacle*k  # larger cost the closer it gets to the obstacle
-        J_fuel = self.heuristic(goal)   # distance from goal
-        J_cost = mu * J_fuel + (1 - mu) * J_threat
-
-        #if J_cost != 0:
-        #    print('J_fuel' + str(J_fuel) + 'J_threat' + str(J_cost))
-        self.fitness = J_cost
-
-    def update_position(self, alpha_position, goal, obstacles):
-        # find new position based on alpha/omega designation
-        if math.sqrt((self.x - goal[0]) ** 2 + (self.y - goal[1]) ** 2) <= MOVEMENT_SPEED:
-            self.path.append(goal)
-            self.temppath.append(goal)
-            self.x = goal[0]
-            self.y = goal[1]
-
-        if self.is_alpha:
-            # calculate distance and direction to goal
-            dx = goal[0] - self.x
-            dy = goal[1] - self.y
-            magnitude = math.sqrt((dx ** 2 + dy ** 2))
-
-            # normalize direction vector to goal
-            if magnitude > 0:
-                dx /= magnitude
-                dy /= magnitude
-
-            # calculate new position based on direction towards goal
-            new_x = self.x + dx * MOVEMENT_SPEED
-            new_y = self.y + dy * MOVEMENT_SPEED
-
-        else:   # omega wolves
-            # implement position update logic based on alpha position
-            strength = random.uniform(0.5, 2)  # randomized strength "pull" towards alpha wolf
-
-            # calculate distance and direction to alpha wolf
-            dx_alpha = alpha_position[0] - self.x
-            dy_alpha = alpha_position[1] - self.y
-            distance_alpha = math.sqrt((dx_alpha ** 2 + dy_alpha ** 2))
-
-            # update position to move towards alpha, dependent on strength variable
-            if distance_alpha > 0:
-                direction_x = int(dx_alpha / distance_alpha * strength * MOVEMENT_SPEED)
-                direction_y = int(dy_alpha / distance_alpha * strength * MOVEMENT_SPEED)
-                new_x = self.x + direction_x
-                new_y = self.y + direction_y
-            else:
-                dx = goal[0] - self.x
-                dy = goal[1] - self.y
-                magnitude = math.sqrt((dx ** 2 + dy ** 2))
-
-                if magnitude > 0:
-                    dx /= magnitude
-                    dy /= magnitude
-
-                new_x = self.x + dx * MOVEMENT_SPEED
-                new_y = self.y + dy * MOVEMENT_SPEED
-
-        # checks and limits new position within search space
-        if self.is_valid((new_x, new_y), obstacles):
-            self.x = new_x
-            self.y = new_y
-            self.path.append((self.x, self.y))
-            self.temppath.append((self.x, self.y,))
-
 
 class Obstacle:
     def __init__(self, x, y, radius):
@@ -221,48 +77,6 @@ class Obstacle:
 
     def draw(self, screen):
         pygame.draw.circle(screen, BLACK, (self.x, self.y), self.radius)
-
-
-class Boundary:
-    def __init__(self, x_i, y_i, x_f, y_f):
-        self.start_x = x_i
-        self.start_y = y_i
-        self.final_x = x_f
-        self.final_y = y_f
-
-
-# ******* new addition below   *******************
-
-# def create_mad_agents(min_x, max_x, min_y, max_y, num_agents, goal, obstacles):
-#     # list to contain all agents
-#     agent_objects = []
-#     #
-#
-#     for agent_id in range(1, num_agents + 1):
-#         x = int(random.uniform(min_x, max_x))
-#         y = int(random.uniform(min_y, max_y))
-#         e_th = .56 * random.uniform(2, max_y)
-#         cool = 0.05 * random.uniform(min_y, max_y)
-#         temp = 1500
-#         count = 50
-#         mad_agent = MADDPG_agent((x, y), goal, count, temp, cool, e_th, obstacles, agent_id)
-#         agent_objects.append(mad_agent)
-#     return agent_objects
-
-
-def create_mad_agents_from_agents(agents_in, goal, obstacles):
-    # list to contain all agents
-    agent_objects = []
-    #
-    for agent in agents_in:
-        # edit parameter initialization at some point
-        # might be randomized later
-        e_th = .8
-        temp = 10
-        mad_agent = MADDPG_agent((agent.x, agent.y), goal, temp, e_th, obstacles, agent.agent_id)
-        agent_objects.append(mad_agent)
-    return agent_objects
-
 
 # ************************************************************
 
@@ -293,6 +107,18 @@ def create_wolf_population(right_x, right_y, num_wolves):
         wolf_objects.append(wolf)
     return wolf_objects
 
+def create_mad_agents_from_agents(agents_in, goal, obstacles):
+    # list to contain all agents
+    agent_objects = []
+    #
+    for agent in agents_in:
+        # edit parameter initialization at some point
+        # might be randomized later
+        e_th = .8
+        temp = 10
+        mad_agent = MADDPG_agent((agent.x, agent.y), goal, temp, e_th, obstacles, agent.agent_id)
+        agent_objects.append(mad_agent)
+    return agent_objects
 
 class Algorithm:
     def __init__(self, list_of_agents, obstacles):
@@ -302,6 +128,8 @@ class Algorithm:
         self.force_record = []
         self.angle_record = [0.0]
 
+
+    # A* search ______________________________________________________________
     def a_star_search(self, goal):
         # Heuristic function (Euclidean distance)
         def heuristic(node, goal_in):
@@ -368,6 +196,7 @@ class Algorithm:
 
         return paths
 
+    # APF search
     def apf_search(self, goal):
         # Compute the attractive force between the agent and the goal
         def attractive_force(agent_pos):
@@ -515,7 +344,7 @@ class Algorithm:
             paths.append(agent.temp_path)
 
         return paths
-
+    # MADDPG search
     def mad_search(self):
         q = []
         y = .54  # discount value, also currently
@@ -606,6 +435,7 @@ class Algorithm:
         print(msg)
         return disp_paths
 
+    # HSGWO search
     def hsgwo_msos(self, goal, max_iterations):
         def heuristic(node, goal):
             x, y = node
